Cambios: 	- Se agrega nuevo parámetro "Nombre índice previo" para indicar el índice previo que se utilizará en caso que sea una partición por términos (esto ahorra el hecho de recrear el corpus y hacer la indexación temporal). En caso que se indique '' se realizará la indexación temporal como se hacía antes. 	- Se transforma a Long la colección en los particionamientos por términos ya que con Integer pinchaba por overflow (probar que se haya solucionado totalmente).

+++ terrier-test/src/main/java/partitioning/IPartitionByTerms.java
@@ -6,6 +6,6 @@ import java.util.Map;
 
 public interface IPartitionByTerms extends IPartitionMethod {
 	
-	public void writeDoc(Map<Integer, Map<Integer, Collection<String>>> mapNodeDocTerm, Map<Integer,String> mapDocDocPath, Integer cantidadCorpus, String destinationFolderPath, List<String> colCorpusTotal);
+	public void writeDoc(Map<Long, Map<Long, Collection<String>>> mapNodeDocTerm, Map<Long,String> mapDocDocPath, Integer cantidadCorpus, String destinationFolderPath, List<String> colCorpusTotal);
 	
 }
